+++ app/src/main/java/nextstep/signup/ui/common/component/SignupTextField.kt
@@ -1,4 +1,4 @@
-package nextstep.signup.ui.component
+package nextstep.signup.ui.common.component
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.fillMaxWidth
